Use `SendMessageTimeout` to send messages to FSUIPC
When FSUIPC is terminated, its window handle is still valid but the
module is not responding to its messages. Thus, using `SendMessage`
may cause that a FSX/P3D plugin that is terminated after FSUIPC is
closed processes a session and never receives a response to the
message.

With this alternative, `SendMessageTimeout` sends a message and returns
if there is no response after given timeout parameter (we are using
10 seconds).